Crate plusDomains1

Features

Handle the following Resources with ease from the central hub ...

activities (get, insert and list)

audiences (list)

circles (add people, get, insert, list, patch, remove, remove people and update)

comments (get, insert and list)

media (insert)

people (get, list, list by activity and list by circle)

Everything else about the plusDomains API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.circles().remove_people(...).doit()
let r = hub.circles().insert(...).doit()
let r = hub.circles().add_people(...).doit()
let r = hub.circles().get(...).doit()
let r = hub.circles().list(...).doit()
let r = hub.circles().update(...).doit()
let r = hub.circles().patch(...).doit()
let r = hub.circles().remove(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
plusDomains1 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate plusDomains1;
use plusDomains1::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use plusDomains1::PlusDomains;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = PlusDomains::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.circles().remove_people("circleId")
             .user_id("et")
             .email("dolores")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Acl
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Activity
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert activities (request|response)

get activities (response)

ActivityActor
The person who performed this activity.

This type is not used in any activity, and only used as part of another schema.

ActivityActorImage
The image representation of the actor.

This type is not used in any activity, and only used as part of another schema.

ActivityActorName
An object representation of the individual components of name.

This type is not used in any activity, and only used as part of another schema.

ActivityFeed
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list activities (response)

ActivityGetMethodBuilder
Get an activity.

ActivityInsertMethodBuilder
Create a new activity for the authenticated user.

ActivityListMethodBuilder
List all of the activities in the specified collection for a particular user.

ActivityMethodsBuilder
A builder providing access to all methods supported on activity resources. It is not used directly, but through the PlusDomains hub.

ActivityObject
The object of this activity.

This type is not used in any activity, and only used as part of another schema.

ActivityObjectActor
If this activity's object is itself another activity, such as when a person reshares an activity, this property specifies the original activity's actor.

This type is not used in any activity, and only used as part of another schema.

ActivityObjectActorImage
The image representation of the original actor.

This type is not used in any activity, and only used as part of another schema.

ActivityObjectAttachments
The media objects attached to this activity.

This type is not used in any activity, and only used as part of another schema.

ActivityObjectAttachmentsEmbed
If the attachment is a video, the embeddable link.

This type is not used in any activity, and only used as part of another schema.

ActivityObjectAttachmentsFullImage
The full image URL for photo attachments.

This type is not used in any activity, and only used as part of another schema.

ActivityObjectAttachmentsImage
The preview image for photos or videos.

This type is not used in any activity, and only used as part of another schema.

ActivityObjectAttachmentsPreviewThumbnails
When previewing, these are the optional thumbnails for the post. When posting an article, choose one by setting the attachment.image.url property. If you don't choose one, one will be chosen for you.

This type is not used in any activity, and only used as part of another schema.

ActivityObjectAttachmentsThumbnails
If the attachment is an album, this property is a list of potential additional thumbnails from the album.

This type is not used in any activity, and only used as part of another schema.

ActivityObjectAttachmentsThumbnailsImage
Image resource.

This type is not used in any activity, and only used as part of another schema.

ActivityObjectPlusoners
People who +1'd this activity.

This type is not used in any activity, and only used as part of another schema.

ActivityObjectReplies
Comments in reply to this activity.

This type is not used in any activity, and only used as part of another schema.

ActivityObjectResharers
People who reshared this activity.

This type is not used in any activity, and only used as part of another schema.

ActivityObjectStatusForViewer
Status of the activity as seen by the viewer.

This type is not used in any activity, and only used as part of another schema.

ActivityProvider
The service provider that initially published this activity.

This type is not used in any activity, and only used as part of another schema.

Audience
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list audiences (none)

AudienceListMethodBuilder
List all of the audiences to which a user can share.

AudienceMethodsBuilder
A builder providing access to all methods supported on audience resources. It is not used directly, but through the PlusDomains hub.

AudiencesFeed
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list audiences (response)

Circle
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

remove people circles (none)

insert circles (request|response)

add people circles (response)

get circles (response)

list circles (none)

update circles (request|response)

patch circles (request|response)

remove circles (none)

CircleAddPeopleMethodBuilder
Add a person to a circle. Google+ limits certain circle operations, including the number of circle adds. Learn More.

CircleFeed
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list circles (response)

CircleGetMethodBuilder
Get a circle.

CircleInsertMethodBuilder
Create a new circle for the authenticated user.

CircleListMethodBuilder
List all of the circles for a user.

CircleMethodsBuilder
A builder providing access to all methods supported on circle resources. It is not used directly, but through the PlusDomains hub.

CirclePatchMethodBuilder
Update a circle's description. This method supports patch semantics.

CirclePeople
The people in this circle.

This type is not used in any activity, and only used as part of another schema.

CircleRemoveMethodBuilder
Delete a circle.

CircleRemovePeopleMethodBuilder
Remove a person from a circle.

CircleUpdateMethodBuilder
Update a circle's description.

Comment
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert comments (request|response)

get comments (response)

list comments (none)

CommentActor
The person who posted this comment.

This type is not used in any activity, and only used as part of another schema.

CommentActorImage
The image representation of this actor.

This type is not used in any activity, and only used as part of another schema.

CommentFeed
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list comments (response)

CommentGetMethodBuilder
Get a comment.

CommentInReplyTo
The activity this comment replied to.

This type is not used in any activity, and only used as part of another schema.

CommentInsertMethodBuilder
Create a new comment in reply to an activity.

CommentListMethodBuilder
List all of the comments for an activity.

CommentMethodsBuilder
A builder providing access to all methods supported on comment resources. It is not used directly, but through the PlusDomains hub.

CommentObject
The object of this comment.

This type is not used in any activity, and only used as part of another schema.

CommentPlusoners
People who +1'd this comment.

This type is not used in any activity, and only used as part of another schema.

Media
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert media (request|response)

MediaAuthor
The person who uploaded this media.

This type is not used in any activity, and only used as part of another schema.

MediaAuthorImage
The author's Google profile image.

This type is not used in any activity, and only used as part of another schema.

MediaExif
Exif information of the media item.

This type is not used in any activity, and only used as part of another schema.

MediaInsertMethodBuilder
Add a new media item to an album. The current upload size limitations are 36MB for a photo and 1GB for a video. Uploads do not count against quota if photos are less than 2048 pixels on their longest side or videos are less than 15 minutes in length.

MediaMethodsBuilder
A builder providing access to all methods supported on media resources. It is not used directly, but through the PlusDomains hub.

PeopleFeed
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list by circle people (response)

list by activity people (response)

list people (response)

PeopleGetMethodBuilder
Get a person's profile.

PeopleListByActivityMethodBuilder
List all of the people in the specified collection for a particular activity.

PeopleListByCircleMethodBuilder
List all of the people who are members of a circle.

PeopleListMethodBuilder
List all of the people in the specified collection.

PeopleMethodsBuilder
A builder providing access to all methods supported on people resources. It is not used directly, but through the PlusDomains hub.

Person
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get people (response)

PersonCover
The cover photo content.

This type is not used in any activity, and only used as part of another schema.

PersonCoverCoverInfo
Extra information about the cover photo.

This type is not used in any activity, and only used as part of another schema.

PersonCoverCoverPhoto
The person's primary cover image.

This type is not used in any activity, and only used as part of another schema.

PersonEmails
A list of email addresses that this person has, including their Google account email address, and the public verified email addresses on their Google+ profile. The plus.profile.emails.read scope is needed to retrieve these email addresses, or the email scope can be used to retrieve just the Google account email address.

This type is not used in any activity, and only used as part of another schema.

PersonImage
The representation of the person's profile photo.

This type is not used in any activity, and only used as part of another schema.

PersonName
An object representation of the individual components of a person's name.

This type is not used in any activity, and only used as part of another schema.

PersonOrganizations
A list of current or past organizations with which this person is associated.

This type is not used in any activity, and only used as part of another schema.

PersonPlacesLived
A list of places where this person has lived.

This type is not used in any activity, and only used as part of another schema.

PersonUrls
A list of URLs for this person.

This type is not used in any activity, and only used as part of another schema.

Place
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

PlaceAddress
The physical address of the place.

This type is not used in any activity, and only used as part of another schema.

PlacePosition
The position of the place.

This type is not used in any activity, and only used as part of another schema.

PlusDomains
Central instance to access all PlusDomains related resource activities

PlusDomainsAclentryResource
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Videostream
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.