Struct plusDomains1::Acl

pub struct Acl {
    pub items: Vec<PlusDomainsAclentryResource>,
    pub kind: Option<String>,
    pub description: Option<String>,
    pub domain_restricted: Option<bool>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

items

The list of access entries.

kind

Identifies this resource as a collection of access controls. Value: "plus#acl".

description

Description of the access granted, suitable for display.

domain_restricted

Whether access is restricted to the domain.